Changes ProductsController#show to use a different type of join.
We were seeing certain products that would display no variants to be
sold which means they couldn't add that product to their cart because
it was using a inner join which meant to be a variant to be sold it
would have to have an option_value. Changing that to a left_outer join
means it shows all variants reguardless if they have an option_value.